tried to make a real component
